requirements.txt
jsonschema
flask
PyYAML
waitress
weasyprint